# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tonezen API
  version: 1.0.0
  description: Offline-first media player API

servers:
  - url: http://localhost:8000/api/v1
    description: Local Kong gateway

security:
  - bearerAuth: []

paths:
  /auth/invite/verify:
    post:
      summary: Verify an invite code before showing signup form
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "200":
          description: Whether the invite code is valid (always 200 to avoid enumeration)
          content:
            application/json:
              schema:
                type: object
                required: [valid]
                properties:
                  valid:
                    type: boolean
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/signup:
    post:
      summary: Register a user with a valid invite code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invite_code, email, password]
              properties:
                invite_code:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 12
                display_name:
                  type: string
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/AuthUser"
        "400":
          description: Invalid signup request or signup failed (generic; does not reveal email existence)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/referral-code:
    get:
      summary: Get current user's referral code
      responses:
        "200":
          description: Referral code
          content:
            application/json:
              schema:
                type: object
                required: [code]
                properties:
                  code:
                    type: string
                    pattern: "^[A-Z0-9]{12}$"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/password/recovery:
    post:
      summary: Send password recovery email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Generic recovery response
          content:
            application/json:
              schema:
                type: object
                required: [sent]
                properties:
                  sent:
                    type: boolean

  /auth/password/update:
    post:
      summary: Set a new password with a recovery access token (password-reset landing only)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [access_token, password]
              properties:
                access_token:
                  type: string
                password:
                  type: string
                  minLength: 12
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                type: object
                required: [updated]
                properties:
                  updated:
                    type: boolean
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid recovery token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/password:
    post:
      summary: Change password while authenticated (requires current password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, password]
              properties:
                current_password:
                  type: string
                password:
                  type: string
                  minLength: 12
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                type: object
                required: [updated]
                properties:
                  updated:
                    type: boolean
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required or current password incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /catalog/cycles:
    get:
      summary: List audiobook cycles
      security: []
      parameters:
        - name: updated_since
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Cycle list
          content:
            application/json:
              schema:
                type: object
                properties:
                  cycles:
                    type: array
                    items:
                      $ref: "#/components/schemas/Cycle"

  /catalog/music:
    get:
      summary: List music albums
      security: []
      parameters:
        - name: updated_since
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Music album list
          content:
            application/json:
              schema:
                type: object
                properties:
                  albums:
                    type: array
                    items:
                      $ref: "#/components/schemas/Book"

  /catalog/books/{bookId}:
    get:
      summary: Get book or album with tracks
      security: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Book detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDetail"
        "404":
          description: Not found

  /downloads/sign:
    post:
      summary: Sign download URLs for tracks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [track_ids]
              properties:
                track_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: Signed URLs
          content:
            application/json:
              schema:
                type: object
                properties:
                  urls:
                    type: array
                    items:
                      type: object
                      properties:
                        track_id:
                          type: string
                          format: uuid
                        url:
                          type: string
                          format: uri

  /progress/audiobooks:
    get:
      summary: Get all audiobook progress for user
      responses:
        "200":
          description: Progress list
          content:
            application/json:
              schema:
                type: object
                properties:
                  progress:
                    type: array
                    items:
                      $ref: "#/components/schemas/AudiobookProgress"

  /progress/audiobooks/{bookId}:
    put:
      summary: Upsert audiobook progress
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [track_id, position_ms, updated_at]
              properties:
                track_id:
                  type: string
                  format: uuid
                position_ms:
                  type: integer
                updated_at:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudiobookProgressUpsertResponse"
        "400":
          description: Invalid request or non-audiobook content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthUser:
      type: object
      required: [id, email]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email

    Cycle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        book_order:
          type: array
          items:
            type: string
        books:
          type: array
          items:
            $ref: "#/components/schemas/Book"

    Book:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        content_type:
          type: string
          enum: [audiobook, music]
        title:
          type: string
        author:
          type: string

    BookDetail:
      allOf:
        - $ref: "#/components/schemas/Book"
        - type: object
          properties:
            tracks:
              type: array
              items:
                $ref: "#/components/schemas/Track"

    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sort_order:
          type: integer
        title:
          type: string
        artist:
          type: string
          nullable: true
        filename:
          type: string
        duration_ms:
          type: integer
        waveform_peaks:
          type: array
          nullable: true
          minItems: 64
          maxItems: 64
          items:
            type: integer
            minimum: 0
            maximum: 100

    AudiobookProgress:
      type: object
      properties:
        book_id:
          type: string
          format: uuid
        track_id:
          type: string
          format: uuid
        position_ms:
          type: integer
        updated_at:
          type: string
          format: date-time

    AudiobookProgressUpsertResponse:
      type: object
      required: [progress]
      properties:
        skipped:
          type: boolean
          description: True when an existing newer server record won LWW conflict resolution.
        progress:
          $ref: "#/components/schemas/AudiobookProgress"

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string